JSON convert failed
userRights changed from int to long

+++ Loxone.Api/Data/LoxApp/MiniserverInfo.cs
@@ -11,7 +11,7 @@ namespace Loxone.Api.Data.LoxApp
         public string Uuid { get; set; }
         public bool isAdmin { get; set; }
         public bool changePassword { get; set; }
-        public int UserRights { get; set; }
+        public long UserRights { get; set; }
     }
     public class MiniserverInfo
     {
